feat: Add support for using jobspec2 parser
This adds in support for using the HCL2 jobspec2 parser with a new -hcl2 flag.
This allows for using new features like `upstreams.datacenter` and terminating
gateways. It should also allow for new features in the future to be used that
are only being added to the new jobspec2 parser.

Signed-off-by: Alex Dulin <alex@morningconsult.com>